analyse_circ_weigh as separate module
also created separate json_circweigh_utils.py module.
