+++ mass_circular_weighing/routines/__init__.py
@@ -1,1 +1,3 @@
-from .run_circ_weigh import check_for_existing_weighdata, check_existing_runs, check_bal_initialised, do_circ_weighing, analyse_weighing
+from .run_circ_weigh import check_existing_runs, check_bal_initialised, do_circ_weighing
+from .json_circweigh_utils import check_for_existing_weighdata
+from .analyse_circ_weigh import analyse_weighing
