+++ mass_circular_weighing/routines/analyse_circ_weigh.py
@@ -1,0 +1,176 @@
+"""
+Functions to analyse circular weighing data
+"""
+from __future__ import annotations
+from typing import TYPE_CHECKING
+
+import numpy as np
+from datetime import datetime
+
+
+from .. import __version__
+from ..log import log
+
+from ..routine_classes.circ_weigh_class import CircWeigh
+from .json_circweigh_utils import *
+from ..constants import SUFFIX, MU_STR, local_backup, IN_DEGREES_C
+
+if TYPE_CHECKING:
+    from mass_circular_weighing.configuration import Configuration
+    from msl.io import JSONWriter
+
+tab = '  '
+
+
+def analyse_weighing(root, url, se, run_id, bal_mode, timed=False, drift=None, EXCL=3, local_backup_folder=local_backup, **metadata):
+    """Analyse a single complete circular weighing measurement using methods in circ_weigh_class
+
+    Parameters
+    ----------
+    root : :class:`root`
+        see msl.io for details
+    url : path
+        path to json file where analysis will be saved (along with measurement run data)
+    se : :class:`str`
+        scheme entry
+    run_id : :class:`str`
+        string in format run_1
+    timed : :class:`bool`, optional
+        if :data:`True`, uses times from weighings, otherwise assumes equally spaced in time
+    drift : :class:`str`, optional
+        set desired drift correction, e.g. 'quadratic drift'.  If :data:`None`, routine selects optimal drift correction
+    EXCL : :class:`float`, optional
+        criterion for excluding a single weighing within an automatic weighing sequence, default set arbitrarily at 3
+    local_backup_folder : path
+        path to local backup folder
+
+    Returns
+    -------
+    :class:`root`
+        the original root object with new analysis data
+    (or None if weighing was not completed)
+    """
+    schemefolder = root['Circular Weighings'][se]
+    weighdata = schemefolder['measurement_' + run_id]
+
+    if not weighdata.metadata.get('Weighing complete'):
+        return None
+
+    log.info(f'CIRCULAR WEIGHING ANALYSIS for scheme entry {se} {run_id}')
+
+    weighing = CircWeigh(se)
+    if timed:
+        times = np.reshape(weighdata[:, :, 0], weighing.num_readings)
+    else:
+        times = []
+    weighing.generate_design_matrices(times)
+
+    if not drift:
+        drift = weighing.determine_drift(weighdata[:, :, 1])  # allows program to select optimum drift correction
+        log.info('Residual std dev. for each drift order:')
+        for key, value in weighing.stdev.items():
+            log.info(tab + key + ':\t' + str(value))
+    else:
+        weighing.determine_drift(weighdata[:, :, 1])
+        log.info('Residual std dev. for '+drift+' correction:\n' + tab
+                 + str(weighing.stdev[drift]))
+
+    massunit = weighdata.metadata.get('Unit')
+    log.info('Selected ' + drift + ' correction (in ' + massunit + ' per ' + weighing.trend + '):\n' + tab
+             + str(weighing.drift_coeffs(drift)))
+
+    analysis = weighing.item_diff(drift)
+
+    log.info('Differences (in ' + massunit + '):')
+    for key, value in weighing.grpdiffs.items():
+        log.info(tab + key + ':\t' + value)
+
+    # print(schemefolder.name + '/analysis_' + run_id)
+    a = root.remove(schemefolder.name+'/analysis_'+run_id)
+
+    weighanalysis = root.require_dataset(schemefolder.name+'/analysis_'+run_id,
+                                                 data=analysis, shape=(weighing.num_wtgrps, 1))
+
+    max_stdev_circweigh = weighdata.metadata.get('Max stdev from CircWeigh ('+MU_STR+'g)')
+
+    analysis_meta = {
+        'Program Version': __version__,
+        'Analysis Timestamp': datetime.now().strftime('%d-%m-%Y %H:%M:%S'),
+        'Residual std devs': str(weighing.stdev),
+        'Selected drift': drift,
+        'Uses mmt times': timed,
+        'Mass unit': massunit,
+        'Drift unit': massunit + ' per ' + weighing.trend,
+        'Acceptance met?': weighing.stdev[drift]*SUFFIX[massunit] < max_stdev_circweigh*SUFFIX['ug'],
+    }
+
+    if bal_mode == 'aw':
+        excl = weighing.stdev[drift]*SUFFIX[massunit] > EXCL*max_stdev_circweigh*SUFFIX['ug']
+    else:
+        excl = not analysis_meta.get('Acceptance met?')
+    analysis_meta['Exclude'] = excl
+
+    for key, value in weighing.driftcoeffs.items():
+        analysis_meta[key] = value
+
+    if not np.isclose(sum(analysis['mass difference']), 0):
+        log.warning('Sum of mass differences is not zero. Analysis not accepted')
+        analysis_meta['Acceptance met?'] = False
+
+    flag = weighdata.metadata.get('Ambient OK?')
+    if not flag:
+        if flag is False:
+            log.warning('Analysis not accepted due to unsuitable ambient conditions during weighing')
+            analysis_meta['Acceptance met?'] = False
+        else:  # e.g. flag is None
+            log.warning('Ambient conditions unavailable during weighing')
+
+    weighanalysis.add_metadata(**analysis_meta)
+
+    timestamp = datetime.strptime(weighdata.metadata.get('Mmt Timestamp'), '%d-%m-%Y %H:%M:%S')
+    save_data(root, url, run_id, timestamp, local_backup_folder)  # save to same file on C: drive as the weighing data
+
+    log.info('Circular weighing analysis for ' + se + ', ' + run_id + ' complete\n')
+
+    return weighanalysis
+
+
+def analyse_old_weighing(cfg, filename, se, run_id):
+    """Analyses a specific weighing run on file, with timed and drift parameters as specified in the configuration"""
+    url = cfg.folder + "\\" + filename + '.json'
+    root = check_for_existing_weighdata(cfg.folder, url, se)
+    weighdata = root['Circular Weighings'][se]['measurement_' + run_id]
+    bal_alias = weighdata.metadata.get('Balance')
+    bal_mode = cfg.equipment[bal_alias].user_defined['weighing_mode']
+    weighanalysis = analyse_weighing(root, url, se, run_id, bal_mode, cfg.timed, cfg.drift)
+
+    return weighanalysis
+
+
+def analyse_all_weighings_in_file(cfg, filename, se):
+    """Analyses all weighings on file for a given scheme entry, with specified timed and drift parameters
+
+    Parameters
+    ----------
+    cfg : :class:`Configuration`
+        configuration instance (see mass_circular_weighing.Configuration)
+    filename : :class:`str`
+        e.g. client_nominal
+    se : :class:`str`
+        scheme entry, as per standard format e.g. "1 1s 0.5+0.5s"
+    """
+    url = cfg.folder + "\\" + filename + '.json'
+
+    i = 1
+    while True:
+        try:
+            run_id = 'run_' + str(i)
+            root = check_for_existing_weighdata(cfg.folder, url, se)
+            weighdata = root['Circular Weighings'][se]['measurement_' + run_id]
+            bal_alias = weighdata.metadata.get('Balance')
+            bal_mode = cfg.equipment[bal_alias].user_defined['weighing_mode']
+            analyse_weighing(root, url, se, run_id, bal_mode, cfg.timed, cfg.drift)
+            i += 1
+        except KeyError:
+            log.info('No more runs to analyse\n')
+            break
